Use files from dianna installation folder in all tutorials
This simplifies the colab setup as we no longer need to download data
files.
The relative paths throughout the tutorials were replaced by paths
starting from the dianna install dir, which should work on all
platforms.
The indentation in the colab setup was also fixed, this was 2 spaces
instead of 4 in many places.